Refactor experiment class
